Redesign crate to guarantee sound deallocation
Relying on the default deallocation of memory by std::boxed::Box is unsound in
that std::alloc::dealloc needs to be called with the same Layout as used for
allocation of the memory. To guarantee sound behavior, we need to implement
our own drop() which deallocates the memory with the correct layout.

To achieve this, we need to entirely change the design of this crate. Instead of
providing traits and their implementations for Box, we now provide a custom
struct AlignedBox that wraps a Box and also stores the layout used for
allocation.

Fixes #1