Wrap Box in mem::ManuallyDrop instead of Option
We need a way to get a raw pointer to the Box's content within Drop::drop(&mut
self) but since we only have a mutable reference to self, the Box needs to be
wrapped in something that provides take(). So far we have used Option for that.

Instead, use mem::ManuallyDrop which is a zero-cost abstraction and semantically
better fits our use case.

See #2