Mark new_uninitialized_sliced as unsafe
Using a slice of elements that are not properly initialized is UB. Although
new_uninitialized_sliced is a private function, mark it as unsafe to make clear
that the caller has the responsibility to initialize all elements of the slice
without looking at the old contents.

For now we need to suppress warnings about the use of unsafe blocks in an
unsafe function. This will be changed based on RFC 2585. Marking a function as
unsafe and using unsafe code within that function should and will be two
different things: https://github.com/rust-lang/rfcs/pull/2585